feat(removecomments): do not remove comments
Enabling "removeComments" causes the TypeScript compiler to not only remove comments but also remove
JSDOC documentation, which is a nuisance when building for npm. This is not really the best way to
remove comments. For projects targeting the web, it makes sense to remove comments as part of a
webpack build. For projects targeting npm, it doesn't make sense to remove comments at all.