Revert "feat(module): compile to ES6 modules instead of CommonJS"
The decision to compile to ES Modules was premature. Node and
webpack tooling isn't quite ready for that yet.

This reverts commit 41623490246c19a76f425d44ac14aeb74052dd93.

BREAKING CHANGE: Modules once again compile to CommonJS, not
ES Modules. This change is backwards-compatible with v2.x but
backwards-incompatible with v3.x and v4.x prereleases. Most
downstream projects should be unaffected.